Solution to Manage Book Library

+++ Manage Book Library@@67877d61dc107486ad5c274c/code.c
@@ -4,9 +4,9 @@ int main(){
     int n;
     scanf("%d",&n);
     char book_name[n][10], book_auth[n][10];
-    int price[n];
+    float price[n];
     for (int i=1;i<=n;i++){
-        scanf("%s %s %d",book_name[i],book_auth[i],&price[i]);
+        scanf("%s %s %f",book_name[i],book_auth[i],&price[i]);
     } 
     int target;
     scanf("%d",&target);
